Bl3 Ogre
Greatly reduces the bullet speed and accuracy of the Ogre in exchange for around 10% dmg 25% fire rate and no spinup time to make it feel like the borderlands 3 ogre. On a non slippery version you will clearly see the rocket trails and smoke (bullet speed comparable to the harold at close range), using a slippery version helps a lot with the slow bullet speed. The weapon becomes a close range beast but hitting mid distance crits/ or moving enemies is much harder.